refactor: refactor code and separate concerns
